refactor noop transform to use dpk_ structures
