+++ transforms/universal/noop/dpk_noop/ray/local_ray.py
@@ -13,9 +13,9 @@
 import os
 import sys
 
-from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.utils import ParamsUtils
-from noop_transform_python import NOOPPythonTransformConfiguration
+from data_processing_ray.runtime.ray import RayTransformLauncher
+from dpk_noop.ray.transform import NOOPRayTransformConfiguration
 
 
 # create parameters
@@ -25,15 +25,20 @@ local_conf = {
     "input_folder": input_folder,
     "output_folder": output_folder,
 }
+worker_options = {"num_cpus": 0.8}
 code_location = {"github": "github", "commit_hash": "12345", "path": "path"}
 params = {
+    # where to run
+    "run_locally": True,
     # Data access. Only required parameters are specified
     "data_local_config": ParamsUtils.convert_to_ast(local_conf),
-    # execution info
+    # orchestrator
+    "runtime_worker_options": ParamsUtils.convert_to_ast(worker_options),
+    "runtime_num_workers": 3,
     "runtime_pipeline_id": "pipeline_id",
     "runtime_job_id": "job_id",
+    "runtime_creation_delay": 0,
     "runtime_code_location": ParamsUtils.convert_to_ast(code_location),
-    "runtime_num_processors": 2,
     # noop params
     "noop_sleep_sec": 1,
 }
@@ -41,6 +46,6 @@ if __name__ == "__main__":
     # Set the simulated command line args
     sys.argv = ParamsUtils.dict_to_req(d=params)
     # create launcher
-    launcher = PythonTransformLauncher(runtime_config=NOOPPythonTransformConfiguration())
+    launcher = RayTransformLauncher(NOOPRayTransformConfiguration())
     # Launch the ray actor(s) to process the input
     launcher.launch()
