+++ transforms/universal/noop/dpk_noop/ray/s3_ray.py
@@ -15,15 +15,21 @@ import sys
 
 from data_processing.utils import ParamsUtils
 from data_processing_ray.runtime.ray import RayTransformLauncher
-from noop_transform_ray import NOOPRayTransformConfiguration
+from dpk_noop.ray.transform import NOOPRayTransformConfiguration
 
 
+print(os.environ)
+# create launcher
+launcher = RayTransformLauncher(NOOPRayTransformConfiguration())
 # create parameters
-input_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "test-data", "input"))
-output_folder = os.path.abspath(os.path.join(os.path.dirname(__file__), "..", "output"))
-local_conf = {
-    "input_folder": input_folder,
-    "output_folder": output_folder,
+s3_cred = {
+    "access_key": "localminioaccesskey",
+    "secret_key": "localminiosecretkey",
+    "url": "http://localhost:9000",
+}
+s3_conf = {
+    "input_folder": "test/noop/input",
+    "output_folder": "test/noop/output",
 }
 worker_options = {"num_cpus": 0.8}
 code_location = {"github": "github", "commit_hash": "12345", "path": "path"}
@@ -31,7 +37,8 @@ params = {
     # where to run
     "run_locally": True,
     # Data access. Only required parameters are specified
-    "data_local_config": ParamsUtils.convert_to_ast(local_conf),
+    "data_s3_cred": ParamsUtils.convert_to_ast(s3_cred),
+    "data_s3_config": ParamsUtils.convert_to_ast(s3_conf),
     # orchestrator
     "runtime_worker_options": ParamsUtils.convert_to_ast(worker_options),
     "runtime_num_workers": 3,
@@ -40,12 +47,11 @@ params = {
     "runtime_creation_delay": 0,
     "runtime_code_location": ParamsUtils.convert_to_ast(code_location),
     # noop params
-    "noop_sleep_sec": 1,
+    "noop_sleep_sec": 5,
 }
-if __name__ == "__main__":
-    # Set the simulated command line args
-    sys.argv = ParamsUtils.dict_to_req(d=params)
-    # create launcher
-    launcher = RayTransformLauncher(NOOPRayTransformConfiguration())
-    # Launch the ray actor(s) to process the input
-    launcher.launch()
+sys.argv = ParamsUtils.dict_to_req(d=params)
+# for arg in sys.argv:
+#     print(arg)
+
+# launch
+launcher.launch()
