--- transforms/universal/noop/ray/.dockerignore
@@ -1,1 +1,0 @@
-venv/
