--- transforms/universal/noop/ray/test-data/expected/metadata.json
@@ -1,46 +1,0 @@
-{
-    "pipeline": "pipeline_id",
-    "job details": {
-        "job category": "preprocessing",
-        "job name": "NOOP",
-        "job type": "ray",
-        "job id": "job_id",
-        "start_time": "2024-03-01 15:17:56",
-        "end_time": "2024-03-01 15:17:57",
-        "status": "success"
-    },
-    "code": [null],
-    "job_input_params": {
-        "sleep": 0,
-        "checkpointing": false,
-        "max_files": -1,
-        "number of workers": 1,
-        "worker options": {
-            "num_cpus": 0.8
-        },
-        "actor creation delay": 0
-    },
-    "execution_stats": {
-        "cpus": 10,
-        "gpus": 0,
-        "memory": 14.031964112073183,
-        "object_store": 2.0
-    },
-    "job_output_stats": {
-        "source_files": 1,
-        "source_size": 16534,
-        "result_files": 1,
-        "result_size": 16534,
-        "table_processing": 0.012392997741699219,
-        "nfiles": 1,
-        "nrows": 5
-    },
-    "source": {
-        "name": "test-data/data_processing/ray/noop/input",
-        "type": "path"
-    },
-    "target": {
-        "name": "/tmp/NOOP4o9gv2bq",
-        "type": "path"
-    }
-}
