--- transforms/universal/noop/spark/.dockerignore
@@ -1,1 +1,0 @@
-venv/
