--- transforms/universal/noop/spark/test-data/expected/metadata.json
@@ -1,46 +1,0 @@
-{
-    "pipeline": "pipeline_id",
-    "job details": {
-        "job category": "preprocessing",
-        "job name": "noop",
-        "job type": "spark",
-        "job id": "job_id",
-        "start_time": "2024-07-31 16:16:13",
-        "end_time": "2024-07-31 16:16:31",
-        "status": "success"
-    },
-    "code": {
-        "github": "github",
-        "commit_hash": "12345",
-        "path": "path"
-    },
-    "job_input_params": {
-        "sleep_sec": 1,
-        "checkpointing": false,
-        "max_files": -1,
-        "random_samples": -1,
-        "files_to_use": [".parquet"]
-    },
-    "execution_stats": {
-        "execution time, min": 0.3105455676714579
-    },
-    "job_output_stats": {
-        "source_size": 753,
-        "result_size": 759,
-        "result_doc_count": 7,
-        "source_files": 1,
-        "result_files": 1,
-        "nrows": 7,
-        "nfiles": 1,
-        "processing_time": 1.0454819202423096,
-        "source_doc_count": 7
-    },
-    "source": {
-        "name": "/Users/borisl/Projects/data-prep-kit/transforms/universal/noop/spark/test-data/input",
-        "type": "path"
-    },
-    "target": {
-        "name": "/Users/borisl/Projects/data-prep-kit/transforms/universal/noop/spark/output",
-        "type": "path"
-    }
-}
