+++ transforms/universal/noop/test/test_noop_ray.py
@@ -12,15 +12,15 @@
 
 import os
 
-from data_processing.runtime.pure_python import PythonTransformLauncher
 from data_processing.test_support.launch.transform_test import (
     AbstractTransformLauncherTest,
 )
-from noop_transform import sleep_cli_param
-from noop_transform_python import NOOPPythonTransformConfiguration
+from data_processing_ray.runtime.ray import RayTransformLauncher
+from dpk_noop.transform import sleep_cli_param
+from dpk_noop.ray.transform import NOOPRayTransformConfiguration
 
 
-class TestPythonNOOPTransform(AbstractTransformLauncherTest):
+class TestRayNOOPTransform(AbstractTransformLauncherTest):
     """
     Extends the super-class to define the test data for the tests defined there.
     The name of this class MUST begin with the word Test so that pytest recognizes it as a test class.
@@ -30,18 +30,18 @@ class TestPythonNOOPTransform(AbstractTransformLauncherTest):
         src_file_dir = os.path.abspath(os.path.dirname(__file__))
         fixtures = []
 
-        launcher = PythonTransformLauncher(NOOPPythonTransformConfiguration())
+        launcher = RayTransformLauncher(NOOPRayTransformConfiguration())
         input_dir = os.path.join(src_file_dir, "../test-data/input")
         expected_dir = os.path.join(src_file_dir, "../test-data/expected")
+        runtime_config = {"run_locally": True}
         transform_config = {sleep_cli_param: 0}
         fixtures.append(
             (
                 launcher,
-                transform_config,
+                transform_config | runtime_config,
                 input_dir,
                 expected_dir,
                 [],  # optional list of column names to ignore in comparing test-generated with expected.
             )
         )
-
         return fixtures
